+++ packages/apollo-client/CHANGELOG.md
@@ -4,6 +4,20 @@
 ### vNEXT
 - Added the graphQLResultHasError in QueryManager.ts to check not only if result.errors is null, but also empty.
 
+### 2.2.4
+- Added `getCacheKey` function to the link context for use in state-link [PR#2998](https://github.com/apollographql/apollo-client/pull/2998)
+
+### 2.2.3
+- dependency updates
+- Provide data when errorPolicy is set "all"
+
+### 2.2.2
+- Fixed potential race condition in mutations
+- Add new fetchPolicy called 'no-cache' to bypass reading from or saving to the cache when making a query
+
+### 2.2.1
+- Allow optional parameter to include queries in standby mode when refetching observed queries [PR#2804](https://github.com/apollographql/apollo-client/pull/2804)
+
 ### 2.2.0
 - include `optimisticResponse` in the context passed to apollo-link for mutations [PR#2704](https://github.com/apollographql/apollo-client/pull/2704)
 - Add cache.writeData to base cache type & DataProxy [PR#2818](https://github.com/apollographql/apollo-client/pull/2818)